+++ PMC-Desktop/CHANGELOG.md
@@ -1,4 +1,58 @@
-#  (2020-07-27)
+#  (2020-07-30)
+
+
+### Bug Fixes
+
+* **UM:** Corrected multiple Direct Reports issues ([c6571cd](https://github.com/PaschalIT/PMC-Desktop/commit/c6571cdd8f092de9549928b5c2385f7006d56c30))
+
+
+### Features
+
+* **UM:** Allow viewing of extraneous users ([ac5521c](https://github.com/PaschalIT/PMC-Desktop/commit/ac5521ce7cd9f441a0d7c03f0b0b49e21d492559))
+* **UM:** Enabled Clear History button ([9e67b08](https://github.com/PaschalIT/PMC-Desktop/commit/9e67b088065b378cb0341cfc1771976b84867c6c))
+
+
+
+## [0.1.6](https://github.com/PaschalIT/PMC-Desktop/compare/0.1.3...0.1.6) (2020-07-29)
+
+
+### Bug Fixes
+
+* **UM:** Manager/DR double click term user fix ([4f23dc8](https://github.com/PaschalIT/PMC-Desktop/commit/4f23dc85362f90798b83935e5581343b36535b3a))
+* **UM:** User History Terminated user ([288f7c3](https://github.com/PaschalIT/PMC-Desktop/commit/288f7c38e1f2dda20035fd6f01651e44904d0835))
+* **UM:** User load crash ([1607ffe](https://github.com/PaschalIT/PMC-Desktop/commit/1607ffe6b5ec5ac8a0c5c98efb9a490368242e6f))
+
+
+### Chore
+
+* **Ver:** Bump to push version ([e237281](https://github.com/PaschalIT/PMC-Desktop/commit/e237281b48397a808c5b3c8a02699fdedb1c8153))
+* **Ver:** Bump to push version 0.1.5 ([e3f5c45](https://github.com/PaschalIT/PMC-Desktop/commit/e3f5c45dbe0ddcdae290dde95c196b0e82045da5))
+
+
+### Documentation
+
+* **Props:** Updated project properties ([7a529e1](https://github.com/PaschalIT/PMC-Desktop/commit/7a529e1464ec633987ef8dd947c8e47a89c59fbc))
+
+
+### Features
+
+* **UM:** Added double click to Manager field ([9a2310f](https://github.com/PaschalIT/PMC-Desktop/commit/9a2310fd2cdd72cdfed5c2c643499e1a4640cbfb))
+* **UM:** Added User History double click ([23fc6d8](https://github.com/PaschalIT/PMC-Desktop/commit/23fc6d857bf8ef137cf13a8b76cf23010af81bd0))
+* **UM:** Added User History tracking ([e53ca70](https://github.com/PaschalIT/PMC-Desktop/commit/e53ca70f4e608a697fdfa21b1abc3a51d73f4171))
+
+
+
+## [0.1.3](https://github.com/PaschalIT/PMC-Desktop/compare/0.1.2-a...0.1.3) (2020-07-27)
+
+
+### Documentation
+
+* **Props:** Updated project properties ([35952f4](https://github.com/PaschalIT/PMC-Desktop/commit/35952f42b518a6d0b4b99717c50a9706df6a64c3))
+
+
+### Features
+
+* **UM:** Added Direct Reports double click ([ea2ff67](https://github.com/PaschalIT/PMC-Desktop/commit/ea2ff67d19122c17414172f45a40e06e62600a0d))
 
 
 
@@ -13,6 +67,11 @@
 ### Documentation
 
 * **Changelog:** Working out changelog styles ([a115752](https://github.com/PaschalIT/PMC-Desktop/commit/a115752f71da35975e2d1818e093c6a16eaa017d))
+
+
+### Features
+
+* **UM:** Added Direct Reports double click ([64c9b3e](https://github.com/PaschalIT/PMC-Desktop/commit/64c9b3eda1d7f74c7503245b383a2ffc27854995))